Move readme file generation code into components directory
