+++ gigadb/app/tools/readme-generator/config/test.php
@@ -17,9 +17,8 @@ return [
         'cache' => [
             'class' => 'yii\caching\FileCache',
         ],
-        // To access DatasetService component in file-worker
-        'datasetService' => [
-            'class' => 'GigaDB\services\DatasetService',
+        'ReadmeGenerator' => [
+            'class' => 'app\components\ReadmeGenerator',
         ],
         'log' => [
             'targets' => [
